// client/src/app/interfaces/projects.d.ts
/*****************************************************************************
 * @Author                : Jbristhuille<jbristhuille@gmail.com>             *
 * @CreatedDate           : 2025-06-06 15:38:23                              *
 * @LastEditors           : Jbristhuille<jbristhuille@gmail.com>             *
 * @LastEditDate          : 2025-06-06 15:38:23                              *
 ****************************************************************************/

export interface IProject {
  id: string,
  name: string,
  apiKey: string,
  createdAt: Date,
  updatedAt: Date
}